gpsd: update to 3.17
This is a substantial rework from the last published version (3.14).
The upstream version now works almost "out of the box", except for one
last-minute bug which is fixed here by a patchfile duplicating the
upstream post-release patch.

Changes from a functional perspective (other than upstream changes):

The universal variant has been disabled, since it didn't work (and
would probably require significant work on the SConstruct script to
implement).

A new "xgps" variant has been added to determine whether the xgps and
xgpsspeed X11 programs are included. This is because they both require
cairo, py-gobject3 and gtk3, expanding recursively to 202 additional
(previously unexpressed) dependencies that aren't otherwise needed.
The fact that these were unconditionally included (but without the
dependencies) created a dilemma for the default, since defaulting to
no means omitting the previously included programs, but defaulting to
yes means adding all the dependencies for programs that may not
actually be used.  The chosen compromise is to default to no but with
a note warning the user about the need for the variant.

An earlier version of this work supported choosing either Qt4 or Qt5
for Qt support.  In the meantime, something changed in Qt5 causing
gpsd not to build with Qt5, so for now the qt variant is back to
Qt4-only.

There's a workaround for the long-standing pkgconfig bug in qt4-mac
(https://trac.macports.org/ticket/50347).

The variants now correctly control the build options and the
dependencies, which wasn't entirely the case previously.

The Python programs are now shebanged to the chosen Python, which for
now is fixed at 2.7.

The test phase is now implemented, running all the provided regression
tests (with a hack to keep them from being unnecessarily slow - see
the comment).

The long description has been expanded and cosmetically cleaned up.

TESTED:
Built and ran on MacPro/10.9, PowerBook/10.5, and VMs for all OS
versions 10.5-10.13.  Tested with no variants, single variants, and
all variants, except that +xgps doesn't currently work on PPC due to
issues with gtk3 and py-gobject3.