app PortGroup : change executable to launch script
by default the app PortGroup places a symlink to
the executable in the MacOS bundle. However, this
symlink fails to launch in many cases, because
the -psn parameter passed by default by MacOS is
rejected by the executable. Change the default
behaviour to making a simple bash launch script
instead, that does not pass -psn.