ntpsec: Fixes build for macOS <10.13.
The 1.1.0 release of ntpsec included further desupport of macOS
such that it no longer builds on 10.12.  This change adds a patch
which reverts the offending changes, as well as making a few
additional changes to allow building on earlier OSX versions.

This code is also visible (for now) at:
     https://gitlab.com/fhgwright/ntpsec/tree/NTPsec_1_1_0_macports
which is derived from:
      https://gitlab.com/fhgwright/ntpsec/tree/NTPsec_1_1_0_fw
by excluding unnecessary changes for this context.

TESTED:
Built and ran on MacPro 10.9, PowerBook 10.5, and VMs for 10.5-10.13.

Closes: https://trac.macports.org/ticket/56087