slatec: set the fortran compiler properly
Setting the Fortran compiler with FC=${configure.fc} in
build.args-append results in an empty FC variable. Oddly, the build
with an empty FC does not fail, but the libslatec.a it produces is
broken. Fix this by setting the build phase explicitly, and increase
the revision number to force the rebuild of broken slatec
installations.

Closes https://trac.macports.org/ticket/57365