pango 1.42.4: Fix large fonts on Retina displays (HiDPI) for quartz
The font rendering with pango 1.42.4 and gtk 3.22.30 shows too large
fonts for all gtk widgets on MacOS HiDPI devices like the MacBookPro
Retina models. The problem is discussed in
https://bugzilla.gnome.org/show_bug.cgi?id=787867
This is a problem between gtk and pango. A fix is already
in the gtk 3.22 and 3.24 branch. However there seem to be no
more releases for 3.22.

The patch here is also applied in the homebrew project. The
patch might be removed with gtk3 3.24.1.